backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
sse-starlette>=1.6.5
openai>=1.0.0
google-genai>=0.3.0
python-dotenv>=1.0.0
